Renames docs manager toolbar and mixin
