+++ modules/@apostrophecms/modal/ui/apos/components/AposDocsManagerToolbar.vue
