+++ modules/@apostrophecms/modal/ui/apos/mixins/AposDocsManagerMixin.js
